Fixed unwanted exception being thrown on body.
When no body is passed, the validation should still be fired. Previous line of code assumed that some sort of body was being passed. Error was 'TypeError: Cannot read property 'fields' of null'. This would hit on the first 'checkBody()` call. With this fix, when no body is passed, the validation returns as intended.